fixup service references after move
