+++ src/viam/sdk/services/private/mlmodel_client.hpp
@@ -18,7 +18,7 @@
 
 #include <grpcpp/channel.h>
 
-#include <viam/sdk/services/mlmodel/mlmodel.hpp>
+#include <viam/sdk/services/mlmodel.hpp>
 
 namespace viam {
 namespace sdk {
