+++ src/viam/sdk/services/private/motion_server.cpp
@@ -6,8 +6,8 @@
 #include <viam/sdk/common/proto_type.hpp>
 #include <viam/sdk/common/service_helper.hpp>
 #include <viam/sdk/common/utils.hpp>
-#include <viam/sdk/services/motion/motion.hpp>
-#include <viam/sdk/services/motion/server.hpp>
+#include <viam/sdk/services/motion.hpp>
+#include <viam/sdk/services/private/motion_server.hpp>
 #include <viam/sdk/spatialmath/geometry.hpp>
 
 namespace viam {
