+++ src/viam/sdk/services/private/proto.hpp
@@ -16,7 +16,7 @@
 
 #include <vector>
 
-#include <viam/sdk/services/mlmodel/mlmodel.hpp>
+#include <viam/sdk/services/mlmodel.hpp>
 
 namespace viam {
 namespace sdk {
