+++ src/viam/sdk/tests/mocks/mock_motion.hpp
@@ -3,7 +3,7 @@
 #include <viam/sdk/common/pose.hpp>
 #include <viam/sdk/common/proto_type.hpp>
 #include <viam/sdk/resource/resource_api.hpp>
-#include <viam/sdk/services/motion/motion.hpp>
+#include <viam/sdk/services/motion.hpp>
 #include <viam/sdk/spatialmath/geometry.hpp>
 
 namespace viam {
